requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
contourpy==1.2.0
cycler==0.12.1
fonttools==4.47.2
idna==3.6
joblib==1.3.2
kiwisolver==1.4.5
matplotlib==3.8.2
numpy==1.26.3
packaging==23.2
pandas==2.2.0
pillow==10.2.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.4
requests==2.31.0
scikit-learn==1.4.0
scipy==1.12.0
seaborn==0.13.2
six==1.16.0
threadpoolctl==3.2.0
tzdata==2023.4
urllib3==2.1.0
